requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
openpyxl==3.1.2
xlrd>=2.0.0
openai>=1.0.0
pathlib2>=2.3.0 